⚡  IMPROVE: extracted - WIP still need testing
