+++ src/__tests__/helpers.test.js
@@ -1,4 +1,4 @@
-const tidyURI = require("../helpers.js");
+const helpers = require("../helpers.js");
 
 test("tidyURI() can tidy URI", () => {
   const site = "sodiumhalogen.com";
@@ -7,18 +7,18 @@ test("tidyURI() can tidy URI", () => {
   const siteHttp3 = "htt://sodiumhalogen.com";
   const siteHttps = "https://sodiumhalogen.com";
 
-  const updatedSite = tidyURI(site);
+  const updatedSite = helpers.tidyURI(site);
   expect(updatedSite).toBe("http://sodiumhalogen.com");
 
-  const updatedSiteHttp = tidyURI(siteHttp);
+  const updatedSiteHttp = helpers.tidyURI(siteHttp);
   expect(updatedSiteHttp).toBe("http://sodiumhalogen.com");
 
-  const updatedSite2 = tidyURI(siteHttp2);
+  const updatedSite2 = helpers.tidyURI(siteHttp2);
   expect(updatedSite2).toBe("http://sodiumhalogen.com");
 
-  const updatedSite3 = tidyURI(siteHttp3);
+  const updatedSite3 = helpers.tidyURI(siteHttp3);
   expect(updatedSite2).toBe("http://sodiumhalogen.com");
 
-  const updatedSiteHttps = tidyURI(siteHttps);
+  const updatedSiteHttps = helpers.tidyURI(siteHttps);
   expect(updatedSiteHttps).toBe("https://sodiumhalogen.com");
 });
